backend/pyproject.toml
[project]
name = "pulse-backend"
version = "1.0.0"
description = "Pulse — Финансово-коммерческая аналитика (backend)"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",

    # Database
    "sqlalchemy>=2.0.36",
    "alembic>=1.13.3",
    "aiosqlite>=0.20.0",

    # HTTP Client (МойСклад API)
    "httpx>=0.27.2",

    # Data Validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",

    # Background Tasks / Scheduling
    "apscheduler>=3.10.4",

    # Utilities
    "python-dateutil>=2.9.0",
    "pytz>=2024.1",

    # Logging
    "structlog>=24.4.0",

    # Rate limiting
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7.0",
    "black>=24.10.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.2",  # для TestClient
    "factory-boy>=3.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# === Ruff (Линтер) ===
[tool.ruff]
target-version = "py312"
line-length = 100
select = ["E", "W", "F", "I", "N", "UP", "B", "A", "C4", "RET", "SIM", "TCH"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"tests/**/*.py" = ["S101"]

# === Black (Форматирование) ===
[tool.black]
target-version = ["py312"]
line-length = 100

# === mypy (Типизация) ===
[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

# === pytest ===
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
